refactor(example): Remove expo and minimize dependencies
Expo is annoying to deal with so let's go back to vanilla RN for the
example app.

One big change is that I am not trying to do any cleverness by only
using one top-level node_modules; this time we let the example app be
entirely self-contained and have it's own dependencies. As long as you
set up the metro config to blacklist the redundant node_modules,
everything works fine.